Make clippy clean, with an allow-list for the port-shape lints
1540 findings. Cargo.toml allows 23 lints whose fix would fight the port —
index loops, wide signatures, C-shaped conditionals, declaration before
assignment — and the other ~1500 are fixed.

Most of it is `cargo clippy --fix` plus applying the machine-applicable
suggestions from the JSON directly. By hand: 12 files had `mod tests`
mid-file with items after it, `manual_clamp` becomes `.clamp(0, 51)`,
`unnecessary_unwrap` becomes let-chains, the bench `transmute`s get their
type annotations, and one nested `format!` folds into `{:.6}`.

Three of the auto-fixes were wrong and are reverted or corrected:
`assert_eq!(x, false)` became `assert!(x)` rather than `assert!(!x)`, which
the test suite caught; a match guard borrowed what its arm mutates
(collapsible_match, now allowed at that one site); and `as_chunks_mut().0`
was iterated with `iter()` where the body needs `iter_mut()`.

Clean on default / wide / scalar / wide+scalar. Byte-identical: diffharness
847/847 over st qp ps ltr sl dl mt def, 701 tests.

+++ rust/crates/openh264-rs/tests/decoder_nodelay_parity_test.rs
@@ -98,13 +98,13 @@ const ROWS: &[Row] = &[
 unsafe fn nodelay_row(data: &[u8]) -> (usize, (i32, i32), String, Vec<i32>, Vec<i32>) {
     unsafe {
         let mut dec: *mut ISVCDecoder = std::ptr::null_mut();
-        assert_eq!(i64::from(WelsCreateDecoder(&mut dec)), CM_RESULT_SUCCESS as i64);
+        assert_eq!(WelsCreateDecoder(&mut dec), CM_RESULT_SUCCESS as i64);
         let mut param = SDecodingParam::default();
         param.uiTargetDqLayer = u8::MAX;
         param.eEcActiveIdc = ERROR_CON_IDC::ERROR_CON_SLICE_COPY;
         param.sVideoProperty.eVideoBsType = VIDEO_BITSTREAM_DEFAULT;
         assert_eq!(
-            i64::from(ISVCDecoder::Initialize(dec, &param as *const SDecodingParam)),
+            ISVCDecoder::Initialize(dec, &param as *const SDecodingParam),
             CM_RESULT_SUCCESS as i64
         );
