+++ rust/crates/openh264-rs/tests/decoder_numref_change_test.rs
@@ -137,13 +137,13 @@ fn num_ref_frame_change_resizes_the_pool_and_matches_the_reference() {
     // split, `ERROR_CON_SLICE_COPY`, one NAL per call, then the end-of-stream drain.
     unsafe {
         let mut decoder: *mut ISVCDecoder = std::ptr::null_mut();
-        assert_eq!(i64::from(WelsCreateDecoder(&mut decoder)), CM_RESULT_SUCCESS as i64);
+        assert_eq!(WelsCreateDecoder(&mut decoder), CM_RESULT_SUCCESS as i64);
         let mut param = SDecodingParam::default();
         param.uiTargetDqLayer = u8::MAX;
         param.eEcActiveIdc = ERROR_CON_IDC::ERROR_CON_SLICE_COPY;
         param.sVideoProperty.eVideoBsType = VIDEO_BITSTREAM_DEFAULT;
         assert_eq!(
-            i64::from(ISVCDecoder::Initialize(decoder, &param as *const SDecodingParam)),
+            ISVCDecoder::Initialize(decoder, &param as *const SDecodingParam),
             CM_RESULT_SUCCESS as i64
         );
