+++ rust/crates/openh264-rs/tests/decoder_parseonly_parity_test.rs
@@ -111,14 +111,14 @@ fn row(call: usize, rv: i32, info: &SParserBsInfo, sha: &str) -> String {
 unsafe fn parseonly_rows(data: &[u8]) -> Vec<String> {
     unsafe {
         let mut dec: *mut ISVCDecoder = std::ptr::null_mut();
-        assert_eq!(i64::from(WelsCreateDecoder(&mut dec)), CM_RESULT_SUCCESS as i64);
+        assert_eq!(WelsCreateDecoder(&mut dec), CM_RESULT_SUCCESS as i64);
         let mut param = SDecodingParam::default();
         param.uiTargetDqLayer = u8::MAX;
         param.eEcActiveIdc = ERROR_CON_IDC::ERROR_CON_SLICE_COPY;
         param.bParseOnly = true;
         param.sVideoProperty.eVideoBsType = VIDEO_BITSTREAM_DEFAULT;
         assert_eq!(
-            i64::from(ISVCDecoder::Initialize(dec, &param as *const SDecodingParam)),
+            ISVCDecoder::Initialize(dec, &param as *const SDecodingParam),
             CM_RESULT_SUCCESS as i64
         );
